10-29 JWT

+++ tensquare_user/src/main/java/com/tensquare/user/service/UserService.java
@@ -1,5 +1,4 @@
 package com.tensquare.user.service;
-
 import com.tensquare.user.dao.UserDao;
 import com.tensquare.user.pojo.User;
 import org.apache.commons.lang3.RandomStringUtils;
@@ -9,13 +8,16 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.jpa.domain.Specification;
 import org.springframework.data.redis.core.RedisTemplate;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 import util.IdWorker;
+import util.JwtUtil;
 
 import javax.persistence.criteria.CriteriaBuilder;
 import javax.persistence.criteria.CriteriaQuery;
 import javax.persistence.criteria.Predicate;
 import javax.persistence.criteria.Root;
+import javax.servlet.http.HttpServletRequest;
 import java.util.*;
 import java.util.concurrent.TimeUnit;
 
@@ -39,6 +41,15 @@ public class UserService {
 
 	@Autowired
 	private RabbitTemplate rabbitTemplate;
+
+	@Autowired
+	private BCryptPasswordEncoder bCryptPasswordEncoder;
+
+	@Autowired
+	private HttpServletRequest request;
+
+	@Autowired
+	private JwtUtil jwtUtil;
 	/**
 	 * 查询全部列表
 	 * @return
@@ -86,6 +97,8 @@ public class UserService {
 	 * @param user
 	 */
 	public void add(User user) {
+
+		user.setPassword(bCryptPasswordEncoder.encode(user.getPassword()));
 		user.setId( idWorker.nextId()+"" );
 		user.setFollowcount(0);//关注数
 		user.setFanscount(0);//粉丝数
@@ -109,6 +122,10 @@ public class UserService {
 	 * @param id
 	 */
 	public void deleteById(String id) {
+		String token = (String) request.getAttribute("claims_admin");
+		if(token == null || "".equals(token)){
+			throw new RuntimeException("权限不足");
+		}
 		userDao.deleteById(id);
 	}
 
@@ -176,8 +193,16 @@ public class UserService {
 		Map<String,String>  map = new HashMap<String,String>();
 		map.put("mobile",mobile);
 		map.put("checkCode",checkCode);
-		rabbitTemplate.convertAndSend("sms",map);
+		//rabbitTemplate.convertAndSend("sms",map);
 		//控制台打印
 		System.out.println("验证码为："+checkCode);
+	}
+
+	public User login(String mobile,String password) {
+		User user = userDao.findBymobile(mobile);
+		if(user != null && bCryptPasswordEncoder.matches(password,user.getPassword())){
+			return user;
+		}
+		return null;
 	}
 }
